% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{footprint_data}
\alias{footprint_data}
\title{Function to produce summarized fishing effort data by gear type and year group while accounting for confidentiality concerns}
\usage{
footprint_data(Year, Longitude, Latitude, Gear_type, Vessel_ID,
  Minimum_vessels = 3, Start_year, Years_to_combine,
  Spatial_resolution = 0.5, Units = "dd", map_projection = NULL)
}
\arguments{
\item{Year}{The year in which the fishing event occurred}

\item{Longitude}{The longitude of the fishing event in decimal degrees}

\item{Latitude}{The latitude of the fishing event in decimal degrees}

\item{Gear_type}{The gear type of the fishing event (e.g. trawl, trap, longline, gillnet)}

\item{Vessel_ID}{The vessel identifier for the set}

\item{Minimum_vessels}{The minimum number of vessels to fish in in a grid cell in order to satisfy confidentiality rules (default = 3)}

\item{Start_year}{The year to start summmarizing data}

\item{Years_to_combine}{The number of years to combine in the summary of data}

\item{Spatial_resolution}{The spatial resolution of the output grid (default = 0.5). If using "ISEA_12" or "ISEA_13" this value should be NA.}

\item{Units}{The units of the spatial resolution, can be "dd" for decimal degrees, "m" for projected data, "ISEA_12" or "ISEA_13"}

\item{map_projection}{The map projection should be specified in proj4 format (e.g. "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs" for decimal degrees or "+proj=utm +zone=59 +ellps=WGS84 +datum=WGS84 +units=m +no_defs" for UTM}
}
\description{
This function takes a raw fishing effort data frame and summarizes it for mapping, creating a footprint_data_object. The raw fishing effort data should 
include rows for each individual fishing event (set), such as a trawl haul or longline set. For each set the year, latitude, longitude, gear type and a 
vessel identifier must be included. A minimum number of vessels can be specified so that the output does not include grid cells where less than this 
number of vessels fished a given gear type in a given year group. The default value is 3 (consistent with Canadian requirements for confidentiality),
The starting year of interest (usually the earliest year in the data) and how many years should be combined should be specified. Additionally,
the spatial resolution that is required and the units of that spatial resolution (defaults are 0.5 decimal degrees of latitude and longitude) need to 
be specified. A map projection (in proj4 format) should also be given.
}
\details{
The function outputs a list of eight data objects that include a data frame of raw fishing effort data (fishing_effort_data) summarized to the spatial 
resolution #' provided for each fishing event with an additional column for the year group to which the event belongs. There should be one record in this 
table for each of the fishing events that were input. This table may not be appropriate to share with the NPFC, as it will contain records for grid cells
where fewer than the minimum number of vessels fished. The second data object output by the function (fishing_footprint_table) is a table that summarizes 
the fishing events by year group, latitude, longitude and gear type. This table produces the number of sets that occurred in each year by each gear type 
at a grid cell with the spatial resolution specified and the center position given by the latitude and longitude where it was grouped. In addition, a column
specifies the number of vessels that are included at each grid cell. The table will not include grid cells where there were fewer than the minimum number of 
vessels fishing. This table may be appropriate for sharing with the wider group, as it will meet confidentiality rules. The third item is a table of the number 
of grid cells by gear type and year group where the minimum number of vessels was not met (not_enough_vessels). This table indicates how many grid cells for which data exists, but
cannot be reported. The fourth item is the year groups identified from the inputs (years) and the fifth item is the gear types represented in the data (gear_
types). The last three items are the spatial resolution and units of the data summary and the map projection of the data (map_projection). These last items
are for tracking inputs and outputs and are needed for creating the figures using the footprint_map function.
}
\examples{
Spatial_resolution_dd<-data.frame(Spatial_resolution=.5,Units="dd")
Start_year<-1991
Years_to_combine<-5
Minimum_vessels<-3
map_projection_dd<-"+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
example_dd<-footprint_data(effort_data$Year,effort_data$Longitude,effort_data$Latitude,effort_data$Gear_type,effort_data$Vessel,Minimum_vessels,Start_year,
Years_to_combine,Spatial_resolution_dd$Spatial_resolution,Spatial_resolution_dd$Units, map_projection=map_projection_dd)
}
\keyword{NPFC,}
\keyword{effort,}
\keyword{fishing}
\keyword{footprint}
